refactor(winston): updated filename
